requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11
nltk==3.8.1
